manus/requirements.txt
numpy>=1.26
pillow>=10.0
scikit-image>=0.22
scipy>=1.11
tqdm>=4.66
ipywidgets>=8.1
matplotlib>=3.8
pytest>=8.0
